requirements.txt
boto3==1.34.131
botocore==1.34.131
aioboto3==13.1.0
aiobotocore==2.13.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
click==8.1.8
fastapi==0.115.6
h11==0.14.0
idna==3.10
passlib==1.7.4
pydantic==2.10.4
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
python-multipart==0.0.6
openai==0.28.0  # 추가된 부분
azure-cognitiveservices-speech==1.41.1
pymongo==4.10.1
openai==0.28.0
elevenlabs==1.50.3

